% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogdownDigest.R
\docType{package}
\name{blogdownDigest-package}
\alias{blogdownDigest-package}
\title{blogdownDiges: A package that extends blogdown to use hash digests to track
changed source files.}
\description{
The blogdownDigest pacakge extends blogdown to use hash digests to track
changed source files. This works more reliably than using file modification
timestamps to figure out which files need to be rebuilt.
}
\section{License}{


 The \pkg{blogdownDigest} package is open source licensed under the
 MIT License.
}

\section{Bug reports}{

\itemize{
 \item blogdownDigest issue tracker (\url{https://github.com/jonathan-g/blogdownDigest/issues})
}
}

