% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{files_to_rebuild}
\alias{files_to_rebuild}
\title{Find Rmd Files that Need Rebuilding.}
\usage{
files_to_rebuild(dir = NULL)
}
\arguments{
\item{dir}{A directory to start in.}
}
\value{
A list of out-of-date RMarkdown files.
}
\description{
Find Rmd Files that Need Rebuilding.
}
